requirements.txt
Flask
psutil
fasttext-wheel